+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## 31-12-2021
     - Re-did title screen as a Menu
     - Re-named renderer methods to stand in defiance of PEP 8 :D 
+    - Created an "EndgameScreen" to display credits, thanks, etc.
 
 
 ## 28-12-2021